requirements.txt
gradio_client
pygame
pyyaml
inquirer
rich